Added package caching

+++ build.sh
@@ -45,7 +45,8 @@ function makedata
     local rmdata="$1"
     local suffix="$2"
     local desc="$3"
-    
+    local curpath="$(pwd)"
+
     echo " -- Building client-side package $1"
     
     pushd "$rmdata.pk3dir"
@@ -54,6 +55,21 @@ function makedata
     echo "   -- Calculating md5 sums"
     find -regex "^\./[^_].*" -type f -exec md5sum '{}' \; > _md5sums
     local sum="$(md5sum "_md5sums" | sed -e 's/ .*//g')"
+    
+    if [ $CACHEPKGS = 1 ] && [ -e "$curpath/pkgcache/$rmdata-$sum.pk3" ]; then
+        echo "   -- A cached package with the same sum already exists, using it"
+        
+        popd
+        cp -v "pkgcache/$rmdata-$sum.pk3" "$NEXDATA/$rmdata-$sum.pk3"
+        echo "   -- Done"
+
+        BUILT_PACKAGES="${BUILT_PACKAGES}$rmdata-$sum.pk3 "
+        BUILT_PKGINFOS="${BUILT_PKGINFOS}_pkginfo_$sum.txt "
+        BUILT_PKGNAMES="${BUILT_PKGNAMES}$1 "
+
+        return
+    fi
+    
     echo "   -- Writing version info"
     echo "RocketMinsta$2 $VERSION client-side package ($3)" >  _pkginfo_$sum.txt
     echo "Built at $BUILD_DATE"                             >> _pkginfo_$sum.txt
@@ -66,6 +82,12 @@ function makedata
     
     echo "   -- Installing to $NEXDATA"
     mv -v "/tmp/$rmdata-${BUILD_DATE_PLAIN}_tmp.zip" "$NEXDATA/$rmdata-$sum.pk3"
+
+    if [ $CACHEPKGS = 1 ]; then
+        echo "   -- Copying the package to cache"
+        cp -v "$NEXDATA/$rmdata-$sum.pk3" pkgcache
+    fi
+
     echo "   -- Done"
     BUILT_PACKAGES="${BUILT_PACKAGES}$rmdata-$sum.pk3 "
     BUILT_PKGINFOS="${BUILT_PKGINFOS}_pkginfo_$sum.txt "
@@ -168,6 +190,17 @@ fi
 if [ -z $IGNOREPKG ]; then
     warn-oldconfig "config.sh" "IGNOREPKG" "(-)"
     IGNOREPKG=(-)
+fi
+
+if [ -z $CACHEPKGS ]; then
+    warn-oldconfig "config.sh" "CACHEPKGS" "0"
+    CACHEPKGS=0
+fi
+
+if [ "$1" = "cleancache" ]; then
+    echo " -- Cleaning package cache"
+    rm -vf pkgcache/*.pk3 || error "rm failed"
+    exit
 fi
 
 if [ "$1" = "release" ]; then
